+++ roseau/load_flow/exceptions.py
@@ -47,6 +47,7 @@ class RoseauLoadFlowExceptionCode(StrEnum):
     BAD_TRANSFORMER_WINDINGS = auto()
     BAD_TRANSFORMER_TYPE = auto()
     BAD_TRANSFORMER_VOLTAGES = auto()
+    BAD_TRANSFORMER_IMPEDANCE = auto()
     BAD_TRANSFORMER_PARAMETERS = auto()
 
     # Switch